+++ core/src/ons/page-loader.js
@@ -18,7 +18,7 @@ import util from 'ons/util';
 import internal from 'ons/internal';
 
 // Default implementation for global PageLoader.
-function loadPage(page, parent, done) {
+function loadPage(page, parent, params, done) {
   internal.getPageHTMLAsync(page).then(html => {
     const element = util.createElement(html.trim());
     parent.appendChild(element);
@@ -54,11 +54,12 @@ export class PageLoader {
 
   /**
    * @param {any} page
-   * @param {Element} parent
+   * @param {Element} parent A location to load page.
+   * @param {Object} params Extra parameters for ons-page.
    * @param {Function} done Take an object that has "element" property and "unload" function.
    */
-  load(page, parent, done) {
-    this._loader(page, parent, result => {
+  load(page, parent, params, done) {
+    this._loader(page, parent, params, result => {
       if (!(result.element instanceof Element)) {
         throw Error('target.element must be an instance of Element.');
       }
@@ -68,8 +69,18 @@ export class PageLoader {
       }
 
       done(result);
-    });
+    }, params);
   }
 }
 
 export const defaultPageLoader = new PageLoader();
+
+export const instantPageLoader = new PageLoader(function(page, parent, params, done) {
+  const element = util.createElement(page.trim());
+  parent.appendChild(element);
+
+  done({
+    element: element,
+    unload: () => element.remove()
+  });
+});