+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 
 v2.0.0-rc.16
 ----
+ * ons-navigator: Added "page" and "pageLoader" property. Changed to use page-loader instead of `ons._internal.getPageHTMLAsync()`.
  * ons-tabbar, ons-tab: Added "page" and "pageLoader" property. Changed to use page-loader instead of `ons._internal.getPageHTMLAsync()`.
  * core: Added ons.defaultPageLoader and ons.PageLoader.
  * ons-page: Changed to accept ".content" and ".background" elements as child contents.